rotech-backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==0.21.0
pandas==2.3.3
numpy==1.26.4
plotly==5.24.1
openpyxl==3.1.5
supabase==2.22.0
openai==2.26.0
python-multipart==0.0.22
